[switch] some a11y improvements
